+++ rendering/surfaces/PCDX11Texture.h
@@ -66,7 +66,7 @@ public:
 
 	// for RenderResource
 	void resFree() override;
-	void resFill(void* src, size_t size, size_t offset) override;
+	void resFill(void* src, uint32_t size, uint32_t offset) override;
 	char *resGetBuffer() override;
 	void resConstruct() override;
 	void resMethod10() override;
